requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-multipart>=0.0.9
cryptography>=43.0.0
Pillow>=10.0.0
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
